Administration tools and bannable Users

+++ app/src/main/kotlin/team/finder/api/teams/TeamsController.kt
@@ -1,42 +1,41 @@
 package team.finder.api.teams
 
-import com.nimbusds.jwt.SignedJWT
-import io.micrometer.core.instrument.Counter
-import io.micrometer.core.instrument.Metrics
-import io.micrometer.core.instrument.Timer
-import org.springframework.data.domain.PageRequest
-import org.springframework.data.domain.Sort
 import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
-import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.web.bind.annotation.*
+import team.finder.api.users.UsersService
 import team.finder.api.utils.AuthUtil
 import java.util.*
 import javax.validation.Valid
 
 @RestController
 @CrossOrigin
-class TeamsController(val service: TeamsService) {
+class TeamsController(
+    val usersService: UsersService,
+    val service: TeamsService
+) {
 
     @GetMapping("/teams")
     fun index(
         @RequestParam(defaultValue = "1") page: Int,
         @RequestParam(defaultValue = "0") skillsetMask: Int,
         @RequestParam(defaultValue = "asc", name = "order") strSortingOption: String,
-    ) : List<Team> {
+    ) : ResponseEntity<Any> {
         val pageIdx = if (page > 0) page else 1
         val boundedSkillsetMask = if (skillsetMask in 1..127) skillsetMask else 0
         val sortType = service.getSortType(strSortingOption)
-        return service.getTeams(pageIdx, boundedSkillsetMask, sortType)
+        return ResponseEntity(service.getTeams(pageIdx, boundedSkillsetMask, sortType), HttpStatus.OK)
     }
 
     @PostMapping("/teams")
     fun add(@Valid @RequestBody teamDto: TeamDto, @RequestHeader(HttpHeaders.AUTHORIZATION) authHeader: String): ResponseEntity<Any> {
+        if (userIsBanned()) return ResponseEntity(HttpStatus.FORBIDDEN)
+
         val userDetails = AuthUtil.getUserDetails()
 
         val authorId: String = userDetails.discordId
-        if (!service.getTeamByAuthorId(authorId).isEmpty) {
+        if (service.getTeamByAuthorId(authorId) != null) {
             // Only one active Team per user
             return ResponseEntity(HttpStatus.CONFLICT)
         }
@@ -45,19 +44,25 @@ class TeamsController(val service: TeamsService) {
         teamDto.authorId = authorId
 
         // Check this author doesn't already have a team before creating one
-        service.createTeam(Team.fromDto(teamDto))
-        return ResponseEntity(HttpStatus.CREATED)
+        val team = service.createTeam(Team.fromDto(teamDto))
+        return ResponseEntity(team, HttpStatus.CREATED)
     }
 
     @GetMapping("/teams/mine")
-    fun view() : Optional<Team> {
+    fun view() : ResponseEntity<Any> {
+        if (userIsBanned()) return ResponseEntity(HttpStatus.FORBIDDEN)
+
         val userDetails = AuthUtil.getUserDetails()
-        return service.getTeamByAuthorId(userDetails.discordId)
+
+        val jsonSerializableTeam = Optional.ofNullable(service.getTeamByAuthorId(userDetails.discordId))
+        return ResponseEntity(jsonSerializableTeam, HttpStatus.OK)
     }
 
     // TODO: Only changed fields
     @PutMapping("/teams/mine")
     fun update(@Valid @RequestBody teamDto: TeamDto, @RequestHeader(HttpHeaders.AUTHORIZATION) authHeader: String) : ResponseEntity<Any> {
+        if (userIsBanned()) return ResponseEntity(HttpStatus.FORBIDDEN)
+
         val userDetails = AuthUtil.getUserDetails()
 
         service.updateTeam(userDetails.discordId, teamDto.description, teamDto.skillsetMask)
@@ -66,9 +71,30 @@ class TeamsController(val service: TeamsService) {
 
     @DeleteMapping("/teams/mine")
     fun delete(@RequestHeader(HttpHeaders.AUTHORIZATION) authHeader: String) : ResponseEntity<Any> {
-        val userDetails = AuthUtil.getUserDetails();
+        if (userIsBanned()) return ResponseEntity(HttpStatus.FORBIDDEN)
+
+        val userDetails = AuthUtil.getUserDetails()
 
         service.deleteTeam(userDetails.discordId)
         return ResponseEntity(HttpStatus.OK)
+    }
+
+    @PostMapping("/teams/report")
+    fun report(@RequestParam("teamId") teamId: Long, @RequestHeader(HttpHeaders.AUTHORIZATION) authHeader: String) : ResponseEntity<Any> {
+        if (userIsBanned()) return ResponseEntity(HttpStatus.FORBIDDEN)
+
+        val team = service.getTeamById(teamId) ?: return ResponseEntity(HttpStatus.NOT_FOUND)
+
+        // TODO: Audit message about this action
+        team.reportCount = team.reportCount + 1
+        service.saveTeam(team)
+
+        return ResponseEntity(HttpStatus.OK)
+    }
+
+    fun userIsBanned() : Boolean {
+        val userDetails = AuthUtil.getUserDetails()
+        val user = usersService.getUser(userDetails.discordId)
+        return user != null && user.isBanned
     }
 }
